requirements.txt
python-telegram-bot==20.0a4
pydantic==1.10.2
requests==2.28.1
python-Levenshtein==0.20.6
